Grid must be calculated on every layout. Not only once on init.
The number of columns for the radio button group grid cannot be
calculated once during the init of the group. Because the number of
columns depends on the number of visible fields. And the visibility can
change at any time. Therefore if no explicit column number is configured
the columns are calculated on every layout.
Furthermore the default layout is vertical again.

Change-Id: I5184dca72f3b01bd04bf5a4f692cf713ed1edb5c
Reviewed-on: https://git.eclipse.org/r/124675
Tested-by: CI Bot
Reviewed-by: Matthias Villiger <mvi@bsi-software.com>