Close & Cancel buttons must ignore enabled_granted property of parents
isEnabledIncludingParents() for close- and cancel-buttons must return
true if the parents are disabled because of the granted dimension only.
This is because in the UI the button is actually enabled to allow to
close a disabled form. Therefore the check to
isEnabledIncludingParents() must return true in that case even though a
parent box is disabled. This only applies if all disabled parent boxes
are disabled because of the granted dimension only!

229670

Change-Id: I87e9fc04f5d9c737872bc2c81ba1d6da14432cf9
Reviewed-on: https://git.eclipse.org/r/124842
Tested-by: CI Bot
Reviewed-by: Matthias Villiger <mvi@bsi-software.com>