Table: ensure cell editing is complete before a key stroke executed
Use case: A key stroke other than "Enter" is defined to save and close
the form.

Pressing that key stroke closes the form without completing the cell
edit, meaning the content of the open cell editor is not saved. When
pressing the button with the mouse, the editor is correctly closed and
the content saved. Reason: There is a mouse down handler which completes
the cell editing when mouse down is pressed, but there is no key
handler.

This works fine for regular value fields but not for cell editors.
Whenever a key stroke or mouse down is pressed, acceptInput() is called
for the value field. The key stroke manager calls that function on the
focused value field. In order to make it work for cell editors, the cell
editor attaches a keyStroke listener and completes the edit whenever a
relevant key stroke is executed.

229031