Table: don't complete cell editing when a key stroke of the popup is pressed
With commit 55769407873ae4932e3e8af5519557f32438e30b completeCellEdit is
called when a key stroke is pressed. This must not be done if the key
stroke belongs to the cell editor itself. Use case: pressing enter has
to close the popup and must not bubble up, otherwise it would close the
form as well. Since the change the new event handler called completeEdit
but did not stop the propagation.