SequenceBox: the box is empty when shown in a context menu
Reason: clone did not consider the fields.

Also added some more clone properties to some other widgets. This
doesn't mean every widget can now be cloned as expected.

LogicalGrid needs to be a new instance otherwise when toggling between
menu-bar and ellipsis menu the logical grid and the gridConfig of the
original were manipulated which resulted in exceptions. Also, if a field
of the cloned box would get invisible, only the grid of the cloned box
would have to be calculated anew.