ViewMenuPopup: fix item height
Calculated height of view menu popup was wrong on certain zoom levels.
Because the height was not set explicitly (only the line-height), the
browser could return fractional values when measuring the items. To fix
this, the 'line-height' was removed and replaced by 'height'. The
content is aligned using the "vertical align helper", so the line-height
property was not necessary anyway.

Change-Id: I69bcea2de339e9bb90858b8905e340810151d931
Reviewed-on: https://git.eclipse.org/r/125089
Tested-by: CI Bot
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>