IFrame: IE does not show scrollbars even though scrollBarEnabled is true
If the IFrame resp. the BrowserField is rendered, the scrollbars are not
shown. This worked with 6.1.

Reason: renderScrollBarEnabled needs to be called before renderLocation.
With 6.1. _renderScrollBarEnabled also called renderLocation. In order
to don't call _renderLocation twice, _renderLocation is only called when
the scrollBarEnabled property changes dynamically. But unfortunately
this broke the initial rendering with IE.

229830