Action: don't execute other actions with same key stroke
Use case: form contains Cancel and Close menu. Both have ESC key stroke.
Pressing ESC executes the first one which will close the form, the
second one must not be executed anymore because the form is already
closed.

The ActionKeyStroke already stopped the propagation of the key event so
that key strokes in another context won't be executed. But key strokes
in the same context will which actually is inconsistent.

-> ActionKeyStroke now set stopImmediatePropagation to true as well.

--------------------------------------------------------------------
Backported from 8.0.x:
- c2733b0e38e1dd25ce319e009bf8eb7d5fe1095c
  Action: don't execute other actions with same key stroke
- 223337bc621f6304ff7e7667d8c531ef55df3b96
  Button/CheckBox: don't execute other buttons with same key stroke
--------------------------------------------------------------------

Change-Id: I88fd2abacfaa60645852868f54c813f5be048c44
Reviewed-on: https://git.eclipse.org/r/139967
Tested-by: CI Bot
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>