Action: don't execute other actions with same key stroke
Use case: form contains Cancel and Close menu. Both have ESC key stroke.
Pressing ESC executes the first one which will close the form, the
second one must not be executed anymore because the form is already
closed.

The ActionKeyStroke already stopped the propagation of the key event so
that key strokes in another context won't be executed. But key strokes
in the same context will which actually is inconsistent.

-> ActionKeyStroke now set stopImmediatePropagation to true as well.