TileGrid: prefSize must not change gridColumnCount
prefSize operations should not modify the state of the widget, resp.
have to revert the changes. In that case it might even create a loop, if
prefSize with a widthHint is called in a similar while loop as done in
TileGridLayout (which is not done by Scout itself but may be done by a
project)