Fix focus() of RadioButtonGroup, SequenceBox, FileChooserButton, Accordion
Also added isFocusable() function to Widget.js. Returns true if the
element returned by getFocusableElement is focusable.

Composite fields now override getFocusableElement, iterate through the
widgets and return the focusable element of the focusable widget. This
replaces the logic in FormField.focus (findFirstFocusableElement) which
did only work in some cases, e.g. for group box because
groupBox.getFocusableElement did not return anything, but did not work
for sequenceBox or radioButtonGroup because the returned $field is not
focusable.

Activate() and focus() now use the same mechanism.