SmartField: setting width on column descriptor does not work correctly
In a regular table, the widths of the columns are used as weights. Since
the proposal chooser of a smart field also uses a regular table, the
behavior should be the same. Unfortunately it does not work, every
column uses the same width as weight.

Reason: the initialWidth is used as weight, but the width was modified
after the table had been created.

Also added some more properties to copy from a column descriptor:
fixedPosition, cssClass, autoOptimizeWidth (currently Scout JS only).
And used the horizontalAlignment of the field as default (also Scout JS
only because the ColumnDescriptor in Java is initialized with -1 which
makes it impossible to see whether this -1 is the default or the
required alignment).

Extracted some methods to make it easier to customize.