DateField: make clear work in touch mode
Even though there is no "x" on a Date Field in touch mode, the
developer could call clear() by himself -> make it work. But: One could
already call setValue(null), which even works if the field is not
rendered. But because clear() seems to be the obvious choice, it should
work as well.