Fix deadly selector
The selector added to SmartField.less destroyed the YUI minifier.
Minifying was not possible anymore and resulted always in a
StackOverflow exception because the created selector was too big to
handle. Instead of 72 selectors for the clearable rules, 1941 selectors
were created.

The selector is over specified anyway which is not necessary and could
decrease performance during runtime.