SmartField: unnecessary right padding when field is disabled
When the smart field is disabled the icon on the right is made invisible
but the space is reserved anyway. If the icon is invisible the text
should be allowed to take the free space.

Also made sure a classic multiline smart field in read only mode has the
same padding as a regular smart field. Currently it uses the same
padding as an alternative field, but other fields don't.

231231