Popup: don't open if parent is detached
If a popup opens while the parent is detached, the content cannot be
rendered, there will likely be exceptions when a child widget of the
popup tries to access the actual html element e.g. using $container[0],
or if it tries to write and read data using $container.data().

One example exception might be this one
Error: Data "scrollbars" missing in $comp doesn't match any elements

Reason: scout.scrollbars.install writes a data
($container.data('scrollbars', scrollbars)) but Scrollbar.js cannot read
it because it is empty due to the detached state. Actually $container
even has length 0 because this.entryPoint() of scout.Popup could not
find any entry point which means the returned entry point was empty.

Implemented same solution as in Tooltip.js (renderLater).

216130