Mobile: startup failure due to a JSON parse exception
Currently it can lead to a JSON parse exception in the following case:
1. Open the site with a mobile browser (e.g. Safari on iOS)
2. Leave the page by entering another URL
3. Return to the page after the session timeout
-> The index.html is loaded from cache and the request for the
models.json is done. The response cannot be parsed because it is the
login site instead of a json based response -> application cannot be
started.

This happens only for applications that load a models.json, which is not
the regular case.

The fix is to prevent returning the login page for requests which expect
json and to reload the page in case a timeout is returned. More details
about the issue and the fix can be found in the source code of App.js

Regarding the fix in ServletFilterHelper:

The ServletFilterHelper already returned a json based session timeout if
the client sent a json request, but only if method is POST. But the
client may also send a GET request and expect JSON which is the case if
models.json is requested.

Also adjusted the content type check: instead of checking the content
type header, the accept header is now checked. This should actually not
change a lot but is actually more precise since the accept header
describes what type of response the client accepts, the content type
header describes what type the request has. But the check is not very
sophisticated, it does not respect the q-factor weighting, it just looks
for a string called application/json in the header. This is sufficient
for our requests.

230707