Allow caching for scout js assets
Texts, models and locales don't use cache headers yet even though they
might get very big (at least texts and models). Unfortunately a finger
print cannot be calculated easily because only the HtmlDocumentParser is
able to do that at the moment, but the assets mentioned above are
defined in index.js and not index.html. That means the cache control
needs to be based on the etag rather than the max age which is still a
lot better than no caching.

Enabling caching for these resources might even fix a bug related to
caching happening on iOS where it can lead to a JSON parse exception in
the following case:
1. Open the site with a mobile browser (e.g. Safari on iOS)
2. Leave the page by entering another URL
3. Return to the page after the session timeout
4. Reload the page, try to login a new
-> The index.html is loaded from cache and the request for the
models.json is done. The response cannot be parsed because it is the
login site instead of a json based response -> application cannot be
started.
Even though it doesn't make a lot sense to me why enabling the caching
should fix it, but I couldn't reproduce the bug with the new settings
anymore. But that actually doesn't mean anything.

This happens only for applications that load a models.json, which is not
the regular case.

230707