Mobile: don't throw already rendered exception when status menu is shown
If status menu mappings are used on a smart field, an already rendered
exception might occur.

Steps to reproduce
1. Open Smart Field form in widgets app in mobile mode
2. Click on default smart field, enter asdf
3. Close the popup, click in the field again -> exception

Also made menus invisible on popup because they don't work anyway. They
are visible on the original field, though.

And: child elements of a modal element are now excluded from glass pane
targets. This was another problem in the above use case: the tooltip was
covered by a glass pane when it was rendered along with the popup.

231370