Dialog disappears if navigation is made invisible
If a dialog is open with display parent = outline, it will disappear if
navigation is made invisible (either programmatically or by pressing the
key stroke). It won't appear again if navigation is made visible again,
only reloading the page resolves the problem.

Solution:
Never use the display parent as parent if it is the outline. The desktop
already has a logic which makes such forms linked to an outline
invisible or visible when the outline changes. But since they are still
linked to the outline they will be removed when the outline is removed,
which must not happen.

Also: glass pane of the navigation has to be re appended after the
navigation is made invisible again. In order to do that the glass pane
renderer now listens for 'render' of the display parent and renders its
glass panes again. To prevent a memory leak in the glass pane renderer,
it also listens for the 'remove' of the individual glass panes and
removes them from the renderer as soon they are removed from the DOM.

222536