Form not unregistered correctly after close if opened as popup window
If a form is opened as popup window, pressing the "x" on the window
closes the form after 2 seconds. This did not work correctly anymore
because form.hide did not call desktop.hideForm anymore because the form
was not rendered at that time. Reason: DesktopFormController removes the
form before calling close, which is a bit strange but obviously
necessary for IE.

Because the popup windows were not cleaned up correctly, exceptions of
type Cannot read property 'popupWindow' of null were likely after a
popup window was opened once.