Outline: scroll position not always correctly restored
If the scroll position of the outline tree is nearly at the bottom,
changing the outline and switching back does not restore the position
correctly. Instead it scrolls to the very end.

The same happens when calling setScrollTop(): Instead of scrolling to
the given position it scrolls to the very end. It doesn't happen always,
it happens if the view port needs to be updated due to the new scroll
position. It seems to be a Chrome bug, it works as expected with IE 11.

Some debugging showed that after reducing the height of the afterFiller
in _renderFiller the scrollTop will be wrong.
Updating the scrollTop in renderFiller or other view range relevant
function is bad because it corrupts smooth scrolling (see also commit
c14ce92e0a7bff568d4f2d715e3061a782e728c2)

Workaround is to render scroll top again after layouting or calling
setScrollTop. Drawback: If scrollTop is set directly on the HTMLElement
it may still scroll wrongly.

Also made sure that clicking on the outline title not only collapses the
outline and removes the selection but also scrolls to the top.