SmartField2: don't show clear icon when loading
When the user enters text while the smart field is loading the clear
icon will be shown. But: it is not layouted correctly and is displayed
on wrong position. Actually the clear icon should not be shown when the
field is loading otherwise the loading state is not visible anymore. 

Also, it may happen that the loading indicator or even the regular icon
is shown at the wrong position. The reason is the same: the icon is not
layouted because at the time FormFieldLayout was called the icon was
invisible. One solution would be to revalidate the layout when loading
changes, but since it is delayed it makes it more complicate than
necessary. A lot easier is to always layout the icon even if it is
invisible. Since the bounds of the icon are not relevant it is possible
even if it is invisible.

Additionally: loading is set to false whenever the user types some text.
This is not intended, the loading indicator should be visible whenever
the field is doing a lookup. The reason is the RemoteLookupCall: it
aborts the previous lookup before doing a new one.

Note: both issues don't happen on branch 8.0. Reasons:
1. The visibility of the clear icon is controlled using CSS and the
selector considers both, has-text and loading.
2. The loading state is not reset because the current lookup is aborted
by the field before the new lookup is scheduled, so the promise won't be
rejected.

229236

Change-Id: I02b4577a717f208f2f5b82817124b71c0ef646c4
Reviewed-on: https://git.eclipse.org/r/126766
Tested-by: CI Bot
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>