DateUtilityTest: Add test for parsing of non-existing time value
Due to daylight-saving time, not all time values are valid for every
date. DateUtility cannot parse these invalid dates as lenient parsing is
disabled. Note that it is also impossible to store such a date in a
java.util.Date exactly, trying to do so shifts the hour value by 1h.

Change-Id: If1db171376eb0242eec9bb43e2e5b7eb863ad61f
Signed-off-by: Jeremie Miserez <jeremie.miserez@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/126757
Tested-by: CI Bot
Reviewed-by: Matthias Otterbach <matthias.otterbach@bsi-software.com>