Table: columns not auto resized on visibility change (ScoutJS)
If visibility of a column changes (e.g. column.setVisible(false)), the
width of the columns need to be automatically adjusted if
autoResizeColumns is set to true.

The is also true for the artificial checkable and row icon columns.
Also removed this.$rows(true).css('width', this.rowWidth); which should
not be necessary because the viewport will be completely rerendered
anyway by _redraw.

Btw: invalidateLayoutTree is necessary if there is no header, otherwise
_rerenderHeaderColumns would trigger it.

229835