Table: add setter for textWrap and multilineText (ScoutJS)
Small drawback: when changing textWrap in Scout Classic, multilineText
will be changed as well and therefore _renderMultilineText is now called
in addition to updateColumnStructure. Since changing these properties on
the fly is a rare case the minor performance loss is acceptable.

Also set columnLayoutDirty to true when toggling auto optimize column
widths in case autoResizeColumns is true and needs to recalculate the
widths.