Fixed wrong color of outline view tab with form as defaultDetailForm
The bug was only visible when the outline was collapsed completely. When
an outline provides a custom form as default detail-form, the color of
the menu-bar of that form is white. Before this bugfix the bg-color of
the view-tab was set to brown, like the main-menubar.

With this bugfix we ensure that the bg-color of the view tab is white
for custom forms.