Widget: restore scrollTop and scrollLeft on render
Until now, scrollTop and left have been restored when a widget was
attached again. This happened for example when switching between views
or tab items. Also, table and tree restored the scrollTop on render.

The problem with that: it is inconsistent. There are other widgets which
would benefit from restoring scrollTop as well (tileGrid, accordion,
group box, etc.). But instead of implementing the same logic as for tree
and table for these specific widgets, every widget now restores the
scroll positions on render. And not only scrollTop but scrollLeft as
well.

This is also another step in removing attach/detach completely.

Implementation notes:
The widget now has a property scrollTop and scrollLeft. To install a
scrollbar, widget._installScrollbars() should be used. That function not
only creates a scrollbar, but also attaches a listener so that changes
of scrollTop and Left can be saved. Uninstalling the scrollbars on
remove is not necessary anymore because the widget takes care of that.
If a widget uses more than one scrollbar it has to deal with them as
before, the new mechanism can only handle one scrollbar per widget,
which is sufficient for 99% of the cases.