Scrollbar is sometimes clipped unnecessarily or not visible at all
It may happen that when setting scrollTop/Left explicitly, the scroll
bar will be clipped or even not visible at all.

Example:
1. Form menu popup with a compact tree is opened
2. The popup is layouted and the scrollTop set in the TreeLayout
3. Setting scrollTop fixes the scrollbar but unfix is not executed yet
due to setTimeout
4. The popup will be repositioned while the scrollbar still has
position: fixed -> Scroll bar is not visible (width: 0) or partially
clipped.

Solution: don't use setTimeout. SetTimeout was introduced mainly for the
table header menu bar, it is not necessary to do it for the scrollbar.
When scrolling using the scrollbar unfix is executed anyway only on
mouse up.

Btw: it is actually a general issue but happens more often on trees since the commit 44fc5c4e644f7cdb19cb8a15b988cebe24762cb5 because renderScrollTop is called double in TreeLayout