TouchSmartField: closing popup while loading does not always work
When the touch popup is open and a slow connection used, removing the
current value (e.g. using x) and touching outside of the popup right
after (while acceptInput is still in progress) the value won't be
removed in the original field.

Reason: beforeClosePopup listens for acceptInput, but beforeClosePopup
won't be called unless acceptInput has been finished. TouchPopup.js
listens for the acceptInput promise and closes the popup only when it
will be resolved. That code was introduced later than the one in
beforeClosePopup, so listening for acceptInput might have been necessary
once but breaks the actual implementation.

Furthermore, SmartField2._isPreventDefaultTabHandling is not used
anywhere -> it can be deleted along with the flag _lookupInProgress.