TileGrid: implement virtual scrolling
Virtual Scrolling of the tile grid is implemented similarly to the one
of Table and Tree. But there are some important differences and
challenges worth pointing out:

Layouting:
The tile grid is based on the logical grid layout, every tile is
positioned absolutely. The table uses a CSS based table layout so adding
a row will increase the scroll height automatically because it has the
correct height and width as soon it is added to the DOM. A tile needs to
be layouted first. This means scrolling cannot just render the viewport
but also has to layout the newly rendered tiles.

Animations:
The tile grid uses a lot of different animations: insert, delete,
filter, sort, move etc. Since most of the tiles are not rendered, the
tiles cannot just be moved from their old position, because there is no
old position.

Filtering:
Previously, not accepted tiles were made invisible. Now, the tiles need
to be removed because the idea is to only render the visible tiles in
the viewport. Since the animation for filtering and removing is not the
same but both need to work with the animateRemoval concept of the
widget, some extensions to the widget had to be made.

Grid Config:
Because it is based on the logical grid, the tiles could have total
different grid configurations, meaning they could have a different size.
Although it could be possible to handle various row heights (as done for
the table) but it would be a lot more complicated because of the
layouting and more expensive too because of the wrapping feature. Also,
since the grid is often used with useUiHeight = true (accordion) the
pref size needs to be correct, which is impossible if the height of the
tiles is not known (e.g. if useUiHeight is true for tiles as well).
Therefore the decision was to only support real grids with tiles of
equal size, which is the main use case.

Virtual vs. non virtual
Because the virtual mode does not support all cases due to the
limitations mentioned above, the non virtual mode needs to still work
well.

Multiple tiles in a row
Compared to the table, a tile grid does not only work with rows, the
individual tiles can be added or removed from a single row. This means
just using _renderViewPort won't work because the tiles could change
inside the view port too. That is why _renderTileDelta was introduced.

A class VirtualScrolling.js was added with the main Virtual Scrolling
feature which are used by Tree and Table, too, with the intention that
maybe it could be shared among these three one time. But at the moment
it is only used by the tile grid.