Table: prevent JS errors on async callbacks
When the table is already destroyed when the table update buffer is
processed, nothing should be done. Similarly, columns should not attempt
to update cell texts when they are already destroyed on callback.

This may happen for example when a smart column is resolved
asynchronously and the lookup call only returns after the form has
already been closed again.

Change-Id: I94733835381ce6b434e71aa4e3bbdd658d3ebf8b