TileGrid Virtual Scrolling: improve performance of non virtual mode
If non virtual mode would be the same as virtual but always using the
max view range, it would be easier to maintain, but the performance
would suffer, especially when working with filtered tiles. Reason: In
virtual mode filtering renders and removes the tiles, this is expensive
if there are a lot of tiles. To make it faster, non virtual mode just
makes them invisible. But of course, this just works to a certain number
of tiles, the virtual mode should be used whenever possible.