Widget: preserving scroll positions decreases performance
When _renderScrollTop or _renderScrollLeft is executed, the current
scroll positions are read from the scrollable element. This forces the
browser to do a reflow which is expensive.

The check was there to detect non scrollable elements and to prevent
updating the scroll position for such elements.

The solution is to set the positions to null initially and don't do
anything while rendering if they are null. Another approach would be to
remember the scrollable on the widget when installing the scrollbars or
to even have a preserveScrollPos property. But at the time
installScrollbars is called, _renderScrollTop may already be executed,
so it would not work that easy.

Also, setScrollTop and setScrollLeft don't fire a property change
anymore. Reason: scrolling did not use the setter so no property change
was fired on scroll (mainly due to fear of poor performance). So it does
not really help if some scrollTop changes trigger an event and some not.
If someone is interested he can always add a scroll listener.