RadioButtonGroup: radio buttons not displayed when loaded async
When using the lookupCall property on RadioButtonGroup, the lookupCall
creates RadioButtons on completion and should insert them into the
RadioButtonGroup. But if this lookupCall takes more time than form
creation etc., the lookupCall returned too late and the RadioButtons
were not rendered, because the RadioButtonGroup was already rendered.
This commit fixes this issue by implementing a similiar functionality as
in GroupBox. With setFields, fields can be dynamically inserted into the
RadioButtonGroup. For better indication, while the lookupCallis
executing, a loading icon is displayed instead.
A new test was added that checks if this.radioButtons equals to
this.fields after creation.

Note: in _renderFields, the layout tree needs to be revalidated to
prevent flickering in the ui (radiobuttons are visible before they are
correctly positioned)

Change-Id: Ic8a9702fc3bca6d50dd7697adfd55402023529c0
Signed-off-by: Samuel Keusch <samuel.keusch@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/127710
Tested-by: CI Bot
Reviewed-by: Andre Wegmueller <awe@bsiag.com>