Table: do not uncheck row on right-click when it's already checked
Applies only to table that have checkableStyle =
scout.Table.CheckableStyle.TABLE_ROW
In this checkable style mode, a row is checked when you click on it. But
when you right-click on an already checked row, it is unchecked and then
the context menu is shown. the desired behavior is, that the row stays
checked and the context menu is shown.


Change-Id: I789e5bd6938e26857c84b5fdb9aefe3ff23cc27b
Signed-off-by: Samuel Keusch <samuel.keusch@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/127791
Tested-by: CI Bot
Reviewed-by: Andre Wegmueller <awe@bsiag.com>