SmartField2: don't perform lookup when field is blurred
Added 'sync' parameter to acceptInput. When aboutToBlurByMouseDown is
called the parameter is set to true. In that case the SmartField does
not perform a search by text lookup, because this would change the order
of events sent to the UI server. Since the clicked field that gains the
focus cannot wait until the lookup is done, the solution is not to
perform a lookup in this case. In every other case the behavior of the
SmartField is the same as before.

228888