Add ValueDo to wrap objects along with their Java type
Sub types of IValueDo provide typed holders for Java objects. This
allows the type to be included in the serialized representation of the
DO and the correct restoration during deserialization. Without those
holders, the target type could only be guessed. Because there is no 1:1
mapping between the serialized representation and the source type, this
guessing cannot be 100% correct.

Example (for JSON serialization):

WITHOUT VALUE DO:
  Java BigDecimal:
    new BigDecimal("12345")
  Plain JSON serialization:
    12345
  Plain deserialization from JSON (best guess):
    new Integer(12345)

WITH VALUE DO:
  Java BigDecimalValueDo:
    BigDecimalValueDo.of(new  BigDecimal("12345"))
  Typed JSON serialization:
    {
      "_type": "BigDecimalValue",
      "value": 12345
    }
Typed deserialization from JSON (no guessing needed):
  BigDecimalValueDo.of(new  BigDecimal("12345"))

Change-Id: If0e4cec9f38bf477593bf9afd39169b5b3c139aa
Reviewed-on: https://git.eclipse.org/r/127981
Tested-by: CI Bot
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>