HtmlDocumentParser: fix external path conversion
Don't use java.io.File to extract the file name from a path. If the path
contains path separators ("/"), they are returned differently on some
operating systems (slash -> backslash). This might lead to broken
results when building the external path.

Instead, just split the string at the last occurrence of "/" and put
everything before as it is into the resulting path.

Change-Id: I9a2ab6b6c159c8eb4c31f3a56277c1139202bd80