TileGrid Virtual Scrolling: support placeholders and Scout Classic
Placeholders are added by the tile grid itself when needed. Currently
they are added as soon as the logical grid is validated. Unfortunately
this is done very ofthe during the layout which would call
_renderTileDelta many times. This makes it very hard to deal with
because the control is lost. So in order to make it easier (and faster)
the preferred size calculation is now completely based on math instead
of looping. This is possible because of the virtual scrolling
restrictions (same sized tiles).

Another solution could be to update the placeholders at the end of the
layouting process rather at the grid validation. If various row heights
need to be supported sometime this could be an option.