Prevent GlobalHttpResourceCache pollution
The GlobalHttpResourceCache is responsible for holding the responses for
valid resource requests (like html, js, css, etc.). As key for the cache
map the path of the resource is used. Such a path may contain a
hex-fingerprint that acts like a version of the file so that it may be
cached by user agents for a very long time because the filename wont
change as long as the content wont change.
Therefore it was possible to request a valid resource (like an existing
css file) with various fingerprints. The server ignored the requested
fingerprint and always found the corresponding resource and returned the
content to the client.
Furthermore the resource was cached on the server so that future
requests to the same resource must not process the file again. This
results in the situation that the same resource is cached again with the
newly requested fingerprint. Using this it was easily possible to bring
the server to memory limit by requesting the same resource with lots of
different fingerprints (cache pollution).
Furthermore the generation of the script used a lot of resources on the
server and could therefore end in starvation and finally in a DoS
situation.

Now the server ignores the fingerprint for the caching which leads to
the situation that after the first request it is always a cache hit
independent of the supplied fingerprint. This solves the pollution of
the cache (memory) and the CPU heavy generation of the script because it
is only done once.
Furthermore the requested fingerprint is compared with the actual
fingerprint of the cached response candidate. In case the fingerprint of
the response would not match the one requested the server will not send
the file but sends a not found status instead. A request to a resource
that does not exist should not return another resource instead.

#230374

Change-Id: Ie0a4643622242065a5f31b319cb21699429917a9
Reviewed-on: https://git.eclipse.org/r/128746
Tested-by: CI Bot
Reviewed-by: Matthias Villiger <mvi@bsi-software.com>