Resolve constants in JSON models.
This allows developers to use a constant-tag in their JSON model, which
is especially useful for enums/constants with numeric values. It's also
easier to find all references for the same constant in .js and .json
files.

Usage: ${const:ON_FIELD}

A widget (like FormField.js) must call the resolveConst function in its
init function in order to resolve known instance variables.