TileGrid Virtual Scrolling: improve tile accordion
One of the big challenges of the virtual scrolling for tiles is the
fact, that the tile grid is often used in combination with useUiHeight.
This is the case if it is used in a tile accordion. This means, the tile
grid itself is not scrollable but the accordion is. So not the tile grid
needs to attach a scroll listener, but the tile accordion resp. the
first scrollable parent. But not only scrolling is relevant, also
collapsing of a group. This will move the groups below up meaning tiles
might need to be rendered. But: the tile grid itself does not know that
it was moved. Unfortunately this cannot be handled in a generic way
because JavaScript does not provide such a listener. So, the tile
accordion needs to take care of that.