Use layoutValidator instead of postRenderActions to set focus deferred
This solves an issue which was found in a case with two modal dialogs A
and B. When dialog A opens another dialog B and dialog B tries to focus
a field during the render process, the focus() function of the field
registers the focus-action as a postRenderAction. A bit later the
_postRender() function of that field is called, postRenderActions are
executed and the field tries to set the focus. But at this point the
FocusContext is not set correctly so the focus-operation fails.

The problem occurs because the Form is still executing render(), but the
focus-context is only set when the _postRender() function of the Form
runs. Because we don't want to change the order of function calls (all
render-calls of Form + children, followed by all postRender-calls of
Form + children), we decided to use the layoutValidator-approach.