TabBox replace attach/detach with render/remove, create Status widget
Status: the field Status is moved to a own widget. So the Status widget
can bre reused from Tabs where a Tab is the selection element of a
TabItem in a TabBox which does not have the FormField as prototype.

TabBox attach/detach: The TabItem got split into two widgets the TabItem
which has the GroupBox as prototype and is the content of a Tab in a
TabBox and the Tab which is the selection element in the TabBox header
to select a Tab. This separation is needed to simplify the render/remove
mechanism replacing the attach/detach.


Change-Id: I37e3c68a0fdba319fc98c8d15687197ea6525482