Add switch theme functionality for Scout JS apps
It is now possible to switch the theme by using desktop.setTheme(name).
The functionality is the same as for Scout classic apps: The theme is
stored in a cookie called scout.ui.theme and then the browser reloaded.
The server will then return the css files for the new theme.