ComposerField: fix memory leak
The memory leak can happen if a form with a composer is created but
never started. Closing a form which has never been started won't execute
dispose() (may be reproduced by calling getSearchForm() during initPage
-> form is created but not started). This behavior of the form is
correct, because dispose() is
the counterpart to init(), so if init() is not called, dispose() won't
be called as well (init will be called by form.start()).

This fact itself won't create a memory leak. The actual problem occurs
because of the composer. The composer creates the root node during
initConfig(). Unfortunately, AbstractTreeNode executes execInit while
the constructor runs (and not during init as every other widget does),
this means it is executed during initConfig() of the composer. Since
execInit of the root node attaches listeners to a possibly global data
model, the listener needs to be removed in execDispose. This is done
correctly, but execDispose may not be called if the form won't be
disposed due to the above reason.

Solved by moving the creation of the root node from initConfig() to
initInernal(). One might argue, that the actual mistake lies in the tree
node, but calling initTreeNode() later instead of during the constructor
could have some unexpected side effects. And since there are no other
known problems regarding this behaviour, it is more safe to let it be.

232371